Threading and Buttons
Added a new thread for DownloadFile, so the program doesn't become
unresponsive while while downloading.
Added functionality to the Download and Upload song menu items.
*****
Currently, the server cannot open the file, and therefore will not send
the file to the client. Can someone look into this?
Finally, we still need the server to open a new socket for the
downloading and uploading.